% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpriceg2plus.R
\name{enforce_monotonicity}
\alias{enforce_monotonicity}
\title{Enforce monotonicity: P(t,T1) >= P(t,T2) for T1 < T2}
\usage{
enforce_monotonicity(bond_prices, maturities)
}
\description{
Enforce monotonicity: P(t,T1) >= P(t,T2) for T1 < T2
}
