% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount_factors.R
\name{esgdiscountfactor}
\alias{esgdiscountfactor}
\title{Stochastic Discount Factors}
\usage{
esgdiscountfactor(r, X)
}
\arguments{
\item{r}{A numeric vector or time series of interest rates}

\item{X}{A numeric vector or time series of cash flows or values to be discounted}
}
\value{
A time series of discounted values
}
\description{
Computes stochastic discount factors or discounted values based on interest rates and cash flows.
}
\details{
This function handles various combinations of scalar and time series inputs for both interest rates and cash flows.
For scalar inputs, it performs simple exponential discounting.
For time series inputs, it computes cumulative discounting over time.
}
\examples{
# Simple scalar discounting
esgdiscountfactor(0.05, 100)

# Time series discounting
#r <- ts(rep(0.05, 10), start = 0, deltat = 1)
#X <- ts(rep(100, 10), start = 0, deltat = 1)
#esgdiscountfactor(r, X)

}
