% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{esgplotts}
\alias{esgplotts}
\title{Plot time series objects}
\usage{
esgplotts(x)
}
\arguments{
\item{x}{A time series object (ts) containing the data to plot}
}
\value{
A ggplot2 object containing the line plot
}
\description{
Creates a line plot of time series data using ggplot2, with each series represented by a different color.
}
\details{
The function converts the time series data into a format suitable for ggplot2 plotting.
Each column of the time series is plotted as a separate line with a different color.
The x-axis represents time/maturity and the y-axis shows the values.
}
\examples{
# Create sample time series data
x <- ts(matrix(rnorm(100), 20, 5), start = 0, deltat = 0.1)

# Plot the time series
esgplotts(x)

}
