% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvjdg2plus.R
\name{rsvjdg2plus}
\alias{rsvjdg2plus}
\title{Simulate SVJD model (with Feller condition check) with G2++ short rates model}
\usage{
rsvjdg2plus(
  n = 10,
  horizon = 5,
  freq = "quarterly",
  S0 = 100,
  rho_stock_vol = -0.7,
  rho_g2plus = -0.99855687,
  ...
)
}
\arguments{
\item{n}{Number of simulations.}

\item{horizon}{Time steps (e.g., 252 for daily data).}

\item{freq}{Frequency ("daily", "weekly", "monthly").}

\item{S0}{Initial stock price (default=100).}

\item{rho_stock_vol}{Correlation stock price / stochastic volatility}

\item{rho_g2plus}{Correlation between G2++ factors}

\item{...}{Additional parameters to be passed to \code{\link{rsvjd}} and
\code{\link{rg2plus}}}
}
\value{
A of time series objects (\code{ts}).
}
\description{
Simulate SVJD model (with Feller condition check) with G2++ short rates model
}
